+++ .vscode/settings.json
@@ -38,7 +38,6 @@
     "local": "github.com/sourcegraph/sourcegraph"
   },
   "jest.jestCommandLine": "pnpm run test",
-  "gulp.autoDetect": "off",
   "npm.packageManager": "pnpm",
   "npm.runSilent": true,
   "typescript.preferences.quoteStyle": "single",
