+++ .vscode/tasks.json
@@ -25,15 +25,6 @@
       "isBackground": true
     },
     {
-      "label": "Watch code generation",
-      "detail": "Watch files and generate types when files are changed",
-      "type": "npm",
-      "script": "watch-generate",
-      "group": "build",
-      "problemMatcher": [],
-      "isBackground": true
-    },
-    {
       "label": "stylelint",
       "command": "pnpm",
       "args": ["run", "lint:css:all", "--", "--formatter", "compact"],
