+++ client/browser/tsconfig.json
@@ -10,7 +10,7 @@
     "rootDir": ".",
     "outDir": "out",
   },
-  "include": ["src", "config", "gulpfile.js", "src/**/*.json", "src/types/**/*.d.ts", "./code-intel-extensions.json"],
+  "include": ["src", "config", "src/**/*.json", "src/types/**/*.d.ts", "./code-intel-extensions.json"],
   "exclude": [
     "out",
     "node_modules",
