+++ client/shared/tsconfig.json
@@ -43,6 +43,6 @@
       "path": "../wildcard",
     },
   ],
-  "include": ["src", "dev", "gulpfile.js", "./src/**/*.json"],
+  "include": ["src", "dev", "./src/**/*.json"],
   "exclude": ["../../node_modules", "./node_modules", "./out", "./src/end-to-end"],
 }
