+++ client/web/dev/esbuild/server.ts
@@ -8,7 +8,13 @@ import signale from 'signale'
 import { STATIC_ASSETS_PATH } from '@sourcegraph/build-config'
 import { buildMonaco } from '@sourcegraph/build-config/src/esbuild/monacoPlugin'
 
-import { ENVIRONMENT_CONFIG, HTTPS_WEB_SERVER_URL, printSuccessBanner } from '../utils'
+import {
+    DEV_SERVER_LISTEN_ADDR,
+    DEV_SERVER_PROXY_TARGET_ADDR,
+    ENVIRONMENT_CONFIG,
+    HTTPS_WEB_SERVER_URL,
+    printSuccessBanner,
+} from '../utils'
 
 import { esbuildBuildOptions } from './config'
 import { assetPathPrefix } from './manifest'
@@ -65,5 +71,31 @@ export const esbuildDevelopmentServer = async (
             printSuccessBanner(['✱ Sourcegraph is really ready now!', `Click here: ${HTTPS_WEB_SERVER_URL}`])
         })
         proxyServer.once('error', error => reject(error))
+    })
+}
+
+if (require.main === module) {
+    async function main(args: string[]): Promise<void> {
+        if (args.length !== 0) {
+            throw new Error('Usage: (no options)')
+        }
+        await esbuildDevelopmentServer(DEV_SERVER_LISTEN_ADDR, app => {
+            app.use(
+                '/',
+                createProxyMiddleware({
+                    target: {
+                        protocol: 'http:',
+                        host: DEV_SERVER_PROXY_TARGET_ADDR.host,
+                        port: DEV_SERVER_PROXY_TARGET_ADDR.port,
+                    },
+                    logLevel: 'error',
+                })
+            )
+        })
+    }
+    // eslint-disable-next-line unicorn/prefer-top-level-await
+    main(process.argv.slice(2)).catch(error => {
+        console.error(error)
+        process.exit(1)
     })
 }
