+++ package.json
@@ -27,8 +27,7 @@
     "build-web": "pnpm --filter @sourcegraph/web run build",
     "build-app-shell": "pnpm --filter @sourcegraph/app-shell run build",
     "watch-web": "pnpm --filter @sourcegraph/web run watch",
-    "generate": "gulp generate",
-    "watch-generate": "gulp watchGenerate",
+    "generate": "pnpm --filter @sourcegraph/shared run generate",
     "test": "jest --testPathIgnorePatterns end-to-end --testPathIgnorePatterns regression --testPathIgnorePatterns integration --testPathIgnorePatterns storybook \"/out/.*test.js\" \"/out/.*test.d.ts\"",
     "_test-integration": "TS_NODE_PROJECT=client/web/src/integration/tsconfig.json mocha --parallel=${CI:-\"false\"} --retries=1 --jobs=2",
     "test-integration": "NODE_ENV=production pnpm _test-integration \"./client/web/src/integration/**/*.test.ts\"",
@@ -146,7 +145,6 @@
     "@types/glob": "7.1.3",
     "@types/google-spreadsheet": "^3.3.1",
     "@types/got": "9.6.11",
-    "@types/gulp": "4.0.7",
     "@types/highlight.js": "9.12.4",
     "@types/history": "^4.7.11",
     "@types/isomorphic-fetch": "^0.0.36",
@@ -225,7 +223,6 @@
     "gql2ts": "^1.10.1",
     "graphql": "^15.4.0",
     "graphql-schema-linter": "^2.0.1",
-    "gulp": "^4.0.2",
     "http-proxy-middleware": "^2.0.6",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^29.7.0",
